requirements.txt
asyncio
openai
python-telegram-bot
requests
yfinance
qrcode
bip32utils
bip39
